+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Version 0.7.0 (2018-02-04)
 
+- Rename ChannelsCount to ChannelCount.
+- Rename SamplesRate to SampleRate.
 - Remove unused ChannelPosition API.
 - Implement Endpoint and Format Enumeration for macos.
 - Implement format handling for macos `build_voice` method.